+++ src/components/games/admin/index.ts
@@ -1,0 +1,2 @@
+export { GameContentTable } from "./GameContentTable";
+export { GameContentForm } from "./GameContentForm";